requirements.txt
# Sea Level Monitoring System - Root Requirements

# ===== CORE BACKEND DEPENDENCIES =====
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.6

# ===== DATABASE & ORM =====
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# ===== DATA PROCESSING =====
pandas==2.2.2
numpy==1.24.3

# ===== CACHING =====
redis==5.0.1

# ===== ENVIRONMENT & CONFIG =====
python-dotenv==1.0.1
pydantic==2.5.3

# ===== MACHINE LEARNING & FORECASTING =====
scikit-learn==1.4.2
statsmodels==0.14.2

# ===== HTTP & REQUESTS =====
httpx==0.27.0
requests==2.31.0

# ===== TESTING =====
pytest==7.4.4
pytest-asyncio==0.23.5

# ===== DEVELOPMENT TOOLS =====
black==24.3.0
flake8==7.0.0

# ===== UTILITIES =====
python-dateutil==2.8.2
pytz==2023.4

# ===== PRODUCTION =====
gunicorn==21.2.0